requirements.txt
tensorflow
streamlit
pillow
numpy
opencv-python==4.8.1
